# swagger_output.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Store Manager",
    "description": "A Node.js/Express.js RESTful API for a store managing system, where one can create, read, update and delete informations about products and sales in a MySQL database.    ",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://lua-store-manager.herokuapp.com/"
    }
  ],
  "paths": {
    "/products/": {
      "get": {
        "tags": [
          "Products"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductList"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "object",
                    "example": "Nome do Produto"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductSingle"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "message": "\\\"name\\\" is required"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "example": {
                  "message": "\\\"name\\\" length must be at least 5 characters long"
                }
              }
            }
          }
        }
      }
    },
...
    "/sales/": {
      "get": {
        "tags": [
          "Sales"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesList"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Sales"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SalesSinglePost"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesCreateResult"
                }
              }
            }
          },
          "400": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "examples": {
                  "productId missing": {
                    "value": {
                      "message": "\\\"productId\\\" is required"
                    }
                  },
                  "quantity missing": {
                    "value": {
                      "message": "\\\"quantity\\\" is required"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": {
                  "message": "Product not found"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "example": {
                  "message": "\\\"quantity\\\" must be greater than or equal to 1"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ProductList": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "number",
              "example": 1
            },
            "name": {
              "type": "string",
              "example": "Nome do produto"
            }
          }
        }
      },
      "ProductSingle": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Nome do produto"
          }
        }
      },
      "SalesList": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "saleId": {
              "type": "number",
              "example": 1
            },
            "date": {
              "type": "string",
              "example": "2021-09-09T04:54:29.000Z"
            },
            "productId": {
              "type": "number",
              "example": 1
            },
            "quantity": {
              "type": "number",
              "example": 2
            }
          }
        }
      },
...
      "SalesSinglePost": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "productId": {
              "type": "number",
              "example": 1
            },
            "quantity": {
              "type": "number",
              "example": 2
            }
          }
        }
      },
      "SalesCreateResult": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "number",
              "example": 3
            },
            "itemsSold": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "productId": {
                    "type": "number",
                    "example": 1
                  },
                  "quantity": {
                    "type": "number",
                    "example": 2
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }
}